pyproject.toml
[tool.poetry]
name = "bottymcbotface"
version = "0.1.0"
description = "A Discord bot for r/Jailbreak server"
authors = ["SlimShadyIAm <me@slim.rocks>"]

[tool.poetry.dependencies]
python = "^3.9.0"
python-dotenv = "^0.15.0"
"discord.py" = "^1.6.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
mongoengine = "^0.21.0"
apscheduler = "^3.6.3"
humanize = "^3.2.0"
pytimeparse = "^1.1.8"
psutil = "^5.7.3"
python-dateutil = "^2.8.1"
fold-to-ascii = "^1.0.2"
PyNaCl = "^1.4.0"
lavalink = "^3.1.2"
spotipy = "^2.16.1"
twemoji-parser = "^0.4.2"
expiringdict = "^1.2.1"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
rope = "^0.18.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
